create Basic Architecture
cmd — This is where the application binaries (i.e. main.go files) live. We’ve also added an inner api sub-folder to allow for multiple binaries which won’t be possible otherwise

db — The db package acts as a bridge between our application and the database. We will also use it later on to store the database migration files

.env — Contains a “key-value” mapping of our environment variables (e.g., the database credentials)

handler — The handler package includes the API route handlers powered by the gin framework

logstash — This is where we keep code related to logstash such as pipeline configurations and the accompanying Dockerfile

models — Models are Golang structs that can be marshaled into the appropriate JSON objects